rename interfaces.ts to hit.ts
