+++ src/main/angular/lab-app/src/app/model/hit.ts
