+++ src/main/angular/lab-app/src/app/services/point.service.ts
@@ -2,7 +2,7 @@ import {Inject, Injectable} from '@angular/core';
 import {AuthService} from "./auth.service";
 import {HttpClient, HttpHeaders} from "@angular/common/http";
 import {Observable} from "rxjs";
-import {Hit} from "../model/interfaces";
+import {Hit} from "../model/hit";
 
 @Injectable({
   providedIn: 'root'
